requirements.txt
altgraph==0.17.4
bottle==0.13.4
cffi==2.0.0
clr_loader==0.2.8
packaging==25.0
pefile==2023.2.7
proxy-tools==0.1.0
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pypresence==4.6.1
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
typing_extensions==4.15.0